[AssetMapper] Split ImportmapManager into 2
