+++ src/Symfony/Component/AssetMapper/ImportMap/ImportMapManager.php
@@ -12,10 +12,8 @@
 namespace Symfony\Component\AssetMapper\ImportMap;
 
 use Symfony\Component\AssetMapper\AssetMapperInterface;
-use Symfony\Component\AssetMapper\Exception\LogicException;
 use Symfony\Component\AssetMapper\ImportMap\Resolver\PackageResolverInterface;
 use Symfony\Component\AssetMapper\MappedAsset;
-use Symfony\Component\AssetMapper\Path\PublicAssetsPathResolverInterface;
 
 /**
  * @author Kévin Dunglas <kevin@dunglas.dev>
@@ -25,12 +23,8 @@ use Symfony\Component\AssetMapper\Path\PublicAssetsPathResolverInterface;
  */
 class ImportMapManager
 {
-    public const IMPORT_MAP_CACHE_FILENAME = 'importmap.json';
-    public const ENTRYPOINT_CACHE_FILENAME_PATTERN = 'entrypoint.%s.json';
-
     public function __construct(
         private readonly AssetMapperInterface $assetMapper,
-        private readonly PublicAssetsPathResolverInterface $assetsPathResolver,
         private readonly ImportMapConfigReader $importMapConfigReader,
         private readonly RemotePackageDownloader $packageDownloader,
         private readonly PackageResolverInterface $resolver,
@@ -67,104 +61,6 @@ class ImportMapManager
     public function update(array $packages = []): array
     {
         return $this->updateImportMapConfig(true, [], [], $packages);
-    }
-
-    public function findRootImportMapEntry(string $moduleName): ?ImportMapEntry
-    {
-        $entries = $this->importMapConfigReader->getEntries();
-
-        return $entries->has($moduleName) ? $entries->get($moduleName) : null;
-    }
-
-    /**
-     * @internal
-     *
-     * @param string[] $entrypointNames
-     *
-     * @return array<string, array{path: string, type: string, preload?: bool}>
-     */
-    public function getImportMapData(array $entrypointNames): array
-    {
-        $rawImportMapData = $this->getRawImportMapData();
-        $finalImportMapData = [];
-        foreach ($entrypointNames as $entrypointName) {
-            $entrypointImports = $this->findEagerEntrypointImports($entrypointName);
-            // Entrypoint modules must be preloaded before their dependencies
-            foreach ([$entrypointName, ...$entrypointImports] as $import) {
-                if (isset($finalImportMapData[$import])) {
-                    continue;
-                }
-
-                // Missing dependency - rely on browser or compilers to warn
-                if (!isset($rawImportMapData[$import])) {
-                    continue;
-                }
-
-                $finalImportMapData[$import] = $rawImportMapData[$import];
-                $finalImportMapData[$import]['preload'] = true;
-                unset($rawImportMapData[$import]);
-            }
-        }
-
-        return array_merge($finalImportMapData, $rawImportMapData);
-    }
-
-    /**
-     * @internal
-     */
-    public function getEntrypointMetadata(string $entrypointName): array
-    {
-        return $this->findEagerEntrypointImports($entrypointName);
-    }
-
-    /**
-     * @internal
-     */
-    public function getEntrypointNames(): array
-    {
-        $rootEntries = $this->importMapConfigReader->getEntries();
-        $entrypointNames = [];
-        foreach ($rootEntries as $entry) {
-            if ($entry->isEntrypoint) {
-                $entrypointNames[] = $entry->importName;
-            }
-        }
-
-        return $entrypointNames;
-    }
-
-    /**
-     * @internal
-     *
-     * @return array<string, array{path: string, type: string}>
-     */
-    public function getRawImportMapData(): array
-    {
-        $dumpedImportMapPath = $this->assetsPathResolver->getPublicFilesystemPath().'/'.self::IMPORT_MAP_CACHE_FILENAME;
-        if (is_file($dumpedImportMapPath)) {
-            return json_decode(file_get_contents($dumpedImportMapPath), true, 512, \JSON_THROW_ON_ERROR);
-        }
-
-        $rootEntries = $this->importMapConfigReader->getEntries();
-        $allEntries = [];
-        foreach ($rootEntries as $rootEntry) {
-            $allEntries[$rootEntry->importName] = $rootEntry;
-            $allEntries = $this->addImplicitEntries($rootEntry, $allEntries, $rootEntries);
-        }
-
-        $rawImportMapData = [];
-        foreach ($allEntries as $entry) {
-            $asset = $this->findAsset($entry->path);
-            if (!$asset) {
-                throw $this->createMissingImportMapAssetException($entry);
-            }
-
-            $path = $asset->publicPath;
-            $data = ['path' => $path, 'type' => $entry->type->value];
-            $rawImportMapData[$entry->importName] = $data;
-        }
-
-        return $rawImportMapData;
     }
 
     /**
@@ -303,112 +199,6 @@ class ImportMapManager
         }
     }
 
-    /**
-     * Adds "implicit" entries to the importmap.
-     *
-     * This recursively searches the dependencies of the given entry
-     * (i.e. it looks for modules imported from other modules)
-     * and adds them to the importmap.
-     *
-     * @param array<string, ImportMapEntry> $currentImportEntries
-     *
-     * @return array<string, ImportMapEntry>
-     */
-    private function addImplicitEntries(ImportMapEntry $entry, array $currentImportEntries, ImportMapEntries $rootEntries): array
-    {
-        // only process import dependencies for JS files
-        if (ImportMapType::JS !== $entry->type) {
-            return $currentImportEntries;
-        }
-
-        if (!$asset = $this->findAsset($entry->path)) {
-            // should only be possible at this point for root importmap.php entries
-            throw $this->createMissingImportMapAssetException($entry);
-        }
-
-        foreach ($asset->getJavaScriptImports() as $javaScriptImport) {
-            $importName = $javaScriptImport->importName;
-
-            if (isset($currentImportEntries[$importName])) {
-                // entry already exists
-                continue;
-            }
-
-            // check if this import requires an automatic importmap entry
-            if ($javaScriptImport->addImplicitlyToImportMap && $javaScriptImport->asset) {
-                $nextEntry = ImportMapEntry::createLocal(
-                    $importName,
-                    ImportMapType::tryFrom($javaScriptImport->asset->publicExtension) ?: ImportMapType::JS,
-                    $javaScriptImport->asset->logicalPath,
-                    false,
-                );
-
-                $currentImportEntries[$importName] = $nextEntry;
-            } else {
-                $nextEntry = $this->findRootImportMapEntry($importName);
-            }
-
-            // unless there was some missing importmap entry, recurse
-            if ($nextEntry) {
-                $currentImportEntries = $this->addImplicitEntries($nextEntry, $currentImportEntries, $rootEntries);
-            }
-        }
-
-        return $currentImportEntries;
-    }
-
-    /**
-     * Given an importmap entry name, finds all the non-lazy module imports in its chain.
-     *
-     * @return array<string> The array of import names
-     */
-    private function findEagerEntrypointImports(string $entryName): array
-    {
-        $dumpedEntrypointPath = $this->assetsPathResolver->getPublicFilesystemPath().'/'.sprintf(self::ENTRYPOINT_CACHE_FILENAME_PATTERN, $entryName);
-        if (is_file($dumpedEntrypointPath)) {
-            return json_decode(file_get_contents($dumpedEntrypointPath), true, 512, \JSON_THROW_ON_ERROR);
-        }
-
-        $rootImportEntries = $this->importMapConfigReader->getEntries();
-        if (!$rootImportEntries->has($entryName)) {
-            throw new \InvalidArgumentException(sprintf('The entrypoint "%s" does not exist in "importmap.php".', $entryName));
-        }
-
-        if (!$rootImportEntries->get($entryName)->isEntrypoint) {
-            throw new \InvalidArgumentException(sprintf('The entrypoint "%s" is not an entry point in "importmap.php". Set "entrypoint" => true to make it available as an entrypoint.', $entryName));
-        }
-
-        if ($rootImportEntries->get($entryName)->isRemotePackage()) {
-            throw new \InvalidArgumentException(sprintf('The entrypoint "%s" is a remote package and cannot be used as an entrypoint.', $entryName));
-        }
-
-        $asset = $this->findAsset($rootImportEntries->get($entryName)->path);
-        if (!$asset) {
-            throw new \InvalidArgumentException(sprintf('The path "%s" of the entrypoint "%s" mentioned in "importmap.php" cannot be found in any asset map paths.', $rootImportEntries->get($entryName)->path, $entryName));
-        }
-
-        return $this->findEagerImports($asset);
-    }
-
-    private function findEagerImports(MappedAsset $asset): array
-    {
-        $dependencies = [];
-        foreach ($asset->getJavaScriptImports() as $javaScriptImport) {
-            if ($javaScriptImport->isLazy) {
-                continue;
-            }
-
-            $dependencies[] = $javaScriptImport->importName;
-
-            // the import is for a MappedAsset? Follow its imports!
-            if ($javaScriptImport->asset) {
-                $dependencies = array_merge($dependencies, $this->findEagerImports($javaScriptImport->asset));
-            }
-        }
-
-        return $dependencies;
-    }
-
     private static function getImportMapTypeFromFilename(string $path): ImportMapType
     {
         return str_ends_with($path, '.css') ? ImportMapType::CSS : ImportMapType::JS;
@@ -428,14 +218,5 @@ class ImportMapManager
         }
 
         return $this->assetMapper->getAssetFromSourcePath($path);
-    }
-
-    private function createMissingImportMapAssetException(ImportMapEntry $entry): \InvalidArgumentException
-    {
-        if ($entry->isRemotePackage()) {
-            throw new LogicException(sprintf('The "%s" vendor asset is missing. Try running the "importmap:install" command.', $entry->importName));
-        }
-
-        throw new LogicException(sprintf('The asset "%s" cannot be found in any asset map paths.', $entry->path));
     }
 }
